+++ src/progress/tests.rs
@@ -295,7 +295,7 @@ mod report {
         let (progress, weak_reporter) = Progress::new(Task::default(), erased_observer);
 
         progress.update(|task| {
-            task.label = Some("label".to_owned());
+            task.label = Some("label".into());
             task.completed = 5;
             task.total = 10;
         });
@@ -305,7 +305,7 @@ mod report {
         let report = reporter.report();
 
         assert_eq!(report.progress_id, progress.id);
-        assert_eq!(report.label.unwrap(), "label".to_owned());
+        assert_eq!(report.label.unwrap(), "label");
         assert_eq!(report.completed, 5);
         assert_eq!(report.total, 10);
         assert_eq!(report.fraction, 0.5);
@@ -321,19 +321,19 @@ mod report {
         let grandchild = Progress::new_with_parent(Task::default(), &child);
 
         parent.update(|task| {
-            task.label = Some("parent".to_owned());
+            task.label = Some("parent".into());
             task.completed = 1;
             task.total = 2;
         });
 
         child.update(|task| {
-            task.label = Some("child".to_owned());
+            task.label = Some("child".into());
             task.completed = 1;
             task.total = 2;
         });
 
         grandchild.update(|task| {
-            task.label = Some("grandchild".to_owned());
+            task.label = Some("grandchild".into());
             task.completed = 1;
             task.total = 2;
         });
@@ -343,7 +343,7 @@ mod report {
         let parent_report = reporter.report();
 
         assert_eq!(parent_report.progress_id, parent.id);
-        assert_eq!(parent_report.label.unwrap(), "parent".to_owned());
+        assert_eq!(parent_report.label.unwrap(), "parent");
         assert_eq!(parent_report.completed, 3);
         assert_eq!(parent_report.total, 6);
         assert_eq!(parent_report.fraction, 0.5);
@@ -352,7 +352,7 @@ mod report {
         let child_report = parent_report.subreports[0].clone();
 
         assert_eq!(child_report.progress_id, child.id);
-        assert_eq!(child_report.label.unwrap(), "child".to_owned());
+        assert_eq!(child_report.label.unwrap(), "child");
         assert_eq!(child_report.completed, 2);
         assert_eq!(child_report.total, 4);
         assert_eq!(child_report.fraction, 0.5);
@@ -361,7 +361,7 @@ mod report {
         let grandchild_report = child_report.subreports[0].clone();
 
         assert_eq!(grandchild_report.progress_id, grandchild.id);
-        assert_eq!(grandchild_report.label.unwrap(), "grandchild".to_owned());
+        assert_eq!(grandchild_report.label.unwrap(), "grandchild");
         assert_eq!(grandchild_report.completed, 1);
         assert_eq!(grandchild_report.total, 2);
         assert_eq!(grandchild_report.fraction, 0.5);